requirements.txt
absl-py==0.7.0
astor==0.7.1
boto3==1.9.112
boto==2.49.0
botocore==1.12.112
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
docutils==0.14
gast==0.2.2
gensim==3.7.1
gitdb2==2.0.5
gitpython==2.1.11
grpcio==1.19.0
h5py==2.9.0
idna==2.8
jmespath==0.9.4
keras-applications==1.0.7
keras-preprocessing==1.0.9
keras==2.2.4
kiwisolver==1.0.1
markdown==3.0.1
matplotlib==3.0.3
mock==2.0.0
numpy==1.16.2
pandas==0.24.1
pbr==5.1.3
protobuf==3.7.0
pymongo==3.7.2
pyparsing==2.3.1
python-dateutil==2.8.0 ; python_version >= '2.7'
pytz==2018.9
pyyaml==3.13
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
smart-open==1.8.0
smmap2==2.0.5
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow==1.13.1
termcolor==1.1.0
urllib3==1.24.1 ; python_version >= '3.4'
werkzeug==0.14.1
wheel==0.33.1 ; python_version >= '3'
